chore: #52 kube secret functionality split
