+++ internal/secret/kubernetes_secret_runner_test.go
@@ -1,4 +1,4 @@
-package cluster
+package secret
 
 import (
 	"fmt"
