+++ internal/secret/secret_manager.go
@@ -2,8 +2,7 @@ package secret
 
 import (
 	"os"
-
-	"github.com/iac-io/myiac/internal/cluster"
+	"github.com/iac-io/myiac/internal/secret"
 	"github.com/iac-io/myiac/internal/commandline"
 )
 
@@ -31,10 +30,10 @@ func NewTlsSecret(name string, tlsCertPath string, tlsKeyPath string) TlsSecret 
 
 type kubernetesSecretManager struct {
 	namespace        string
-	kubernetesRunner cluster.KubernetesRunner
+	kubernetesRunner KubernetesSecretRunner
 }
 
-func NewKubernetesSecretManager(namespace string, kubernetesRunner cluster.KubernetesRunner) SecretManager {
+func NewKubernetesSecretManager(namespace string, kubernetesRunner secret.KubernetesRunner) SecretManager {
 	return &kubernetesSecretManager{
 		namespace:        namespace,
 		kubernetesRunner: kubernetesRunner,
@@ -42,11 +41,10 @@ func NewKubernetesSecretManager(namespace string, kubernetesRunner cluster.Kuber
 }
 
 func CreateKubernetesSecretManager(namespace string) SecretManager {
-	return NewKubernetesSecretManager(namespace, cluster.NewKubernetesRunner(commandline.NewEmpty()))
+	return NewKubernetesSecretManager(namespace, secret.NewKubernetesRunner(commandline.NewEmpty()))
 }
 
 func (ksm kubernetesSecretManager) CreateTlsSecret(secret TlsSecret) {
-
 	_ = os.Rename(secret.tlsKeyPath, tlsKeyPathTmp)
 	_ = os.Rename(secret.tlsCertPath, tlsCertPathTmp)
 	ksm.kubernetesRunner.CreateTlsSecret(secret.name, ksm.namespace, tlsKeyPathTmp, tlsCertPathTmp)
@@ -58,4 +56,8 @@ func (ksm kubernetesSecretManager) FindTlsSecret(secretName string) {
 
 func (ksm kubernetesSecretManager) CreateFileSecret(secretName string, filePath string) {
 	ksm.kubernetesRunner.CreateFileSecret(secretName, ksm.namespace, filePath)
+}
+
+func (ksm kubernetesSecretManager) CreateLiteralSecret(secretName string, literalsMap map[string]string) {
+	ksm.kubernetesRunner.CreateLiteralSecret(secretName, ksm.namespace, literalsMap)
 }
